ai-worker/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
openai==1.76.0
httpx==0.28.1
supabase==2.15.2
Pillow==11.2.1
python-dotenv==1.1.0
av==13.1.0
ffmpeg-python==0.2.0
python-multipart==0.0.20
